task2/requirements.txt
pandas
openpyxl
numpy
python-dateutil
PyMuPDF
faiss-cpu
sentence-transformers
streamlit
google-genai
